add iceberg code example
